requirements.txt
requests>=2.31.0
matplotlib>=3.7.0
seaborn>=0.12.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
aiofiles>=23.0.0
jinja2>=3.1.0
numpy>=1.24.0